refac: lottoNumbers의 타입 변환(List -> Set)
이외에 convertMoneyToLottosSize() 함수의 위치 변경, lottoNumbersString의
네이밍 변경
