+++ src/main/java/lotto/domain/LottoComparator.java
@@ -1,9 +1,10 @@
 package lotto.domain;
 
 import java.util.List;
+import java.util.Set;
 
 class LottoComparator {
-    static int compare(Lotto lotto, List<Integer> winningNumbers) {
+    static int compare(Lotto lotto, Set<Integer> winningNumbers) {
         return lotto.matchWinningNumbers(winningNumbers);
     }
 }
