+++ src/main/java/lotto/domain/LottoConverter.java
@@ -1,0 +1,8 @@
+package lotto.domain;
+
+public class LottoConverter {
+    private static final int PRICE = 1_000;
+    static int convertMoneyToLottosSize(int money) {
+        return money / PRICE;
+    }
+}
