+++ src/test/java/lotto/domain/RankTest.java
@@ -11,8 +11,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import com.sun.org.apache.bcel.internal.generic.FALOAD;
-
 class RankTest {
 
     @ParameterizedTest
